Fixes for these issues:
1 - Although it might be hard, I still think that after creating a new
group or contact it should be visible in the list and selected.

2 - When adding a new contact can it default to the selected group, if
any?

3 - If I create a new contact and add them to the current group, the
contact list for that group does not update after I click Save. You have
to click the group name again to reload.

4 - Similarly, if I edit a contact and remove them from the current
group the contact list for that group does not update after I click
Save. You have to click the group name again to reload.

5 - If I am adding or editing a contact and click a group or contact on
the left, the add/edit screen is still visible.  Not sure if that's
intentional or not. Might be ok.